Separate table main component, small saga fixes
